chore: implemented ai suggestions
