+++ static/scripts/teams.js
@@ -2,7 +2,7 @@
 // jshint esversion: 6
 
 import 'jquery-datetimepicker';
-import { archiveDownload, humanReadableFileSize } from './download';
+import archiveDownload from './download';
 import './jquery/datetimepicker-easy';
 import { initVideoconferencing } from './videoconference';
 
@@ -56,6 +56,39 @@ function getSelectedFileIds() {
 
 function downloadFiles(selectedFileIds = []) {
 	archiveDownload(getDownloadRequestBody(), selectedFileIds);
+}
+
+function humanReadableFileSize(originalFilesize) {
+	const bytesToMbytes = 1024 * 1024;
+	const mb = originalFilesize / bytesToMbytes;
+	const options = { minimumFractionDigits: 2, maximumFractionDigits: 2 };
+	const formated = new Intl.NumberFormat('de-DE', options).format(mb);
+
+	const result = `${formated} MB`;
+
+	return result;
+}
+
+function sumFileSizes(fileIds) {
+	let totalSize = 0;
+
+	fileIds.forEach((fileId) => {
+		const fileSize = Number.parseInt($(`.filetree-file input[data-file-id="${fileId}"]`).data('file-size'), 10);
+		totalSize += fileSize;
+	});
+
+	return totalSize;
+}
+
+function updateFilesizeSum() {
+	const selectedFileIds = getSelectedFileIds();
+	const totalSize = sumFileSizes(selectedFileIds);
+	const humanReadableSize = humanReadableFileSize(totalSize);
+
+	$('.filesize-sum .current-total').text(humanReadableSize);
+
+	const areAnyFilesSelected = selectedFileIds.length > 0;
+	$('.btn-file-selective-download-submit').prop('disabled', !areAnyFilesSelected);
 }
 
 $(document).ready(() => {
@@ -195,40 +228,27 @@ $(document).ready(() => {
 		$('.files-grid').hide();
 		$('.directories').hide();
 		$('.button-list').hide();
+		updateFilesizeSum();
 		$('html, body').animate({
 			scrollTop: $('.selective-download-filelist').offset().top,
 		}, 800);
 	});
-
-	function sumFileSizes(fileIds) {
-		let totalSize = 0;
-
-		fileIds.forEach((fileId) => {
-			const fileSize = Number.parseInt($(`.filetree-file input[data-file-id="${fileId}"]`).data('file-size'), 10);
-			totalSize += fileSize;
-		});
-
-		return totalSize;
-	}
-
-	function updateFilesizeSum() {
-		const selectedFileIds = getSelectedFileIds();
-		const totalSize = sumFileSizes(selectedFileIds);
-		const humanReadableSize = humanReadableFileSize(totalSize);
-
-		$('.filesize-sum .current-total').text(humanReadableSize);
-
-		const areAnyFilesSelected = selectedFileIds.length > 0;
-		$('.btn-file-selective-download-submit').prop('disabled', !areAnyFilesSelected);
-	}
 
 	$('.folder-checkbox').click((e) => {
 		e.stopPropagation();
 		const isChecked = $(e.currentTarget).is(':checked');
 		const folderId = $(e.currentTarget).data('file-id');
 
-		// Check/uncheck all child checkboxes
 		$(`.filetree-file input[data-file-parent-id="${folderId}"]`).prop('checked', isChecked);
+		const toggleDescendants = (parentId) => {
+			$(`.filetree-file input[data-file-parent-id="${parentId}"]`).prop('checked', isChecked);
+			$(`.folder-checkbox[data-file-parent-id="${parentId}"]`).each((_, el) => {
+				$(el).prop('checked', isChecked);
+				toggleDescendants($(el).data('file-id'));
+			});
+		};
+
+		toggleDescendants(folderId);
 
 		updateFilesizeSum();
 	});
